refactor: 수익률 계산 정확도를 위해 double 사용

+++ src/main/kotlin/lotto/constant/OutputMessage.kt
@@ -18,7 +18,7 @@ enum class OutputMessage(val message: String) {
         println(message)
     }
 
-    fun display(ratio: Float) {
+    fun display(ratio: Double) {
         println(message.format(ratio))
     }
 }
